% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_by_sd.R
\name{cat_by_sd}
\alias{cat_by_sd}
\title{Categorize a vector according to standard deviation values}
\usage{
cat_by_sd(x, n = 1, categories = NULL, na.resp = NA)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{A number that multiplies the standard deviation.}

\item{categories}{Categories used to classify the numeric elements in x. It must be of length 3. Default is NULL.}

\item{na.resp}{Response when \code{NA} is found. Default is \code{NA}.}
}
\value{
A character vector containing the categories used to classify the values of the original vector based on the quantiles defined by the proportionality of the standard error.
}
\description{
Categorize a vector according to standard deviation values
}
\examples{
set.seed(123)
s <- rnorm(20)

cat_by_sd(s)

cat_by_sd(s, categories = c("low", "mid", "high"))

cat_by_sd(s, n = 0.5)

s[1:5] <- NA

cat_by_sd(s)

cat_by_sd(s, na.resp = "nope")
}
