% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_intervals.R
\name{cor_intervals}
\alias{cor_intervals}
\title{Frequency of sliced correlation range.}
\usage{
cor_intervals(x, absolute = FALSE, inter = 0.2)
}
\arguments{
\item{x}{A correlation matrix.}

\item{absolute}{A boolean value (\code{True} or \code{FALSE}). If \code{FALSE} (default) returns a correlation interval between -1 and 1. If \code{TRUE} computes absolute correlations and therefore returns a range between 0 and 1.}

\item{inter}{A value (0 - 0.5] used to cut the correlation interval. Default is 0.2.}
}
\value{
A data frame with the frequencies and rates of the sliced correlation range.
}
\description{
Frequency of sliced correlation range.
}
\examples{
corr_mtcars <- cor(mtcars)

cor_intervals(corr_mtcars)

cor_intervals(corr_mtcars, absolute = TRUE, inter = 0.1)

}
