% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_run_helpers.R
\name{get_asset_id}
\alias{get_asset_id}
\title{Get the Earth Engine system ID of an asset}
\usage{
get_asset_id(asset)
}
\arguments{
\item{asset}{An \code{ee$ImageCollection} or a single \code{ee$Image}.}
}
\value{
A character scalar with the asset's system ID, or \code{NULL} if
it could not be retrieved.
}
\description{
Retrieves the \code{"system:id"} property of an \code{ee$ImageCollection}
(using its first image) or of a single \code{ee$Image}. Used internally
to identify which known satellite/product an asset belongs to (see
\code{\link{get_bands}} and \code{\link{get_native_scale}}).
}
