% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_helpers.R
\name{git_stash_pop}
\alias{git_stash_pop}
\title{Restore (Pop) a Stashed Set of Changes}
\usage{
git_stash_pop(stash_ref = "stash@{0}")
}
\arguments{
\item{stash_ref}{Character. Which stash entry to restore, in \code{git
stash} reference syntax. Default \code{"stash@{0}"}, the most recently
created entry. Use \code{\link{git_stash_list}} to find other entries.}
}
\value{
Invisibly returns the integer exit status of \code{git stash pop}
(\code{0} on success).
}
\description{
Wrapper around \code{git stash pop}. Mentioned as the recovery step after
\code{\link{git_pull_force_remote}} discards local changes into a stash
via \code{\link{git_stash_save}}.
}
\examples{
\dontrun{
git_stash_pop()
git_stash_pop("stash@{1}")
}
}
\seealso{
\code{\link{git_stash_list}}, \code{\link{git_stash_save}}
}
