% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_map.R
\name{permute_map}
\alias{permute_map}
\title{Permute a double-input function operation}
\usage{
permute_map(vec_list_df, .f = `-`, ...)
}
\arguments{
\item{vec_list_df}{A vector, list or data frame.}

\item{.f}{A double input-function.}

\item{...}{Additional arguments for the double-input function (if needed).}
}
\value{
Returns a list with \code{n * (n - 1)} elements resulting from the double-input function operation (defined in \code{.f}) permuted on the \code{n} elements of the object defined in \code{vec_list_df}.
}
\description{
Permute a double-input function operation
}
\details{
This function should be used especially in cases where the input order of the arguments of the double-input function defined in \code{.f} matters (ex. \code{-}).
\itemize{
\item Functions such as \code{cor()} and \code{sum()} are not good options (input order does not matter);
\item Dot-dot-dot functions are also allowed (ex. \code{rbind()}, \code{cbind()}, etc);
\item If possible, functions with a formula input can be modified to a double-input function (\verb{see examples}).
}
}
\examples{
# Load imports
library(dplyr)

# Operation on a vector (default: `-`)
c(a = 1, b = 2, c = 3) |>
  permute_map()

# Operation on a list (difference between elements)
list(a = 1:5, b = 4:10, c = 3:5) |>
  permute_map(setdiff)

# Operation on a data frame (linear model)
lm_quad <- function(y, x) {
  lm(y ~ x + I(x^2))
}

mtcars |>
  select_at(vars(disp:drat)) |>
  permute_map(lm_quad)
}
\seealso{
\code{\link[=comb_map]{comb_map()}} for combining instead of permuting.
}
