% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.R
\name{redundancy}
\alias{redundancy}
\title{Find redundant variables in a correlation matrix}
\usage{
redundancy(x, inf_limit = 1)
}
\arguments{
\item{x}{A correlation matrix of a given data frame.}

\item{inf_limit}{A numeric value indicating the correlation coefficient threshold (default = 1).}
}
\value{
A data frame with columns 'Var1', 'Var2', and 'Correlation' containing pairs of redundant variables and their correlation coefficients.
}
\description{
This function returns the pairs of redundant variables defined by those that are greater than or equal to the inf_limit.
}
\examples{
# Load packages
library(dplyr)

# Creating a data frame with highly correlated variables.
set.seed(123)
mtcars |>
    select_at(vars(disp:qsec)) |>
    tibble::add_column(mtcars |>
        select_at(vars(disp:qsec)) |>
        mutate_all(~. + rnorm(1, sd = 1))) |>
    cor() -> corr

# Finding redundancy
corr |> redundancy()

corr |> redundancy(inf_limit = 0.95)

}
