% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_all_but_diag.R
\name{select_all_but_diag}
\alias{select_all_but_diag}
\title{Remove the diagonal from a matrix and collapse the remaining elements}
\usage{
select_all_but_diag(x, by = "row")
}
\arguments{
\item{x}{a square matrix.}

\item{by}{accepts only "row" or "col".}
}
\value{
Returns a non-square matrix of dimension (n - 1 x n) or (n x n - 1) if \code{by} is respectively set to "row" or "col".
}
\description{
This function removes the diagonal from a square matrix and collapses the remaining elements.
}
\examples{
# Defining a matrix object
set.seed(123)
m <- matrix(sample(1:9), 3, 3)
m

select_all_but_diag(x = m, by = "row")

select_all_but_diag(x = m, by = "col")
}
